changed used summonerIds from int to expected long.

+++ RiotGamesAPIWrapper/HechTex.RiotGamesAPIWrapper/APICaller.cs
@@ -65,7 +65,7 @@ namespace HechTex.RiotGamesAPIWrapper
         /// <param name="region">Region to search in.</param>
         /// <param name="summonerId">The summoner's id.</param>
         /// <returns>List of RunePages.</returns>
-        internal IList<RunePage> GetRunePages(Regions region, int summonerId)
+        internal IList<RunePage> GetRunePages(Regions region, long summonerId)
         {
             var urlsegs = new Dictionary<string, string> {
                 { "region", GetRegionString(region) }, { "summonerId", summonerId.ToString() } };
